Move kafka to tensorflow_io/core
