--- tensorflow_io/kafka/BUILD
@@ -1,29 +1,0 @@
-licenses(["notice"])  # Apache 2.0
-
-package(default_visibility = ["//visibility:public"])
-
-load(
-    "//:tools/build/tensorflow_io.bzl",
-    "tf_io_copts",
-)
-
-cc_library(
-    name = "kafka_ops",
-    srcs = [
-        "kernels/kafka_dataset_ops.cc",
-        "kernels/kafka_kernels.cc",
-        "kernels/kafka_layer_kernels.cc",
-        "kernels/kafka_sequence.cc",
-        "ops/dataset_ops.cc",
-        "ops/kafka_ops.cc",
-    ],
-    copts = tf_io_copts(),
-    linkstatic = True,
-    deps = [
-        "//tensorflow_io/core:dataset_ops",
-        "//tensorflow_io/core:sequence_ops",
-        "@avro",
-        "@kafka",
-    ],
-    alwayslink = 1,
-)
